Add support for form reset buttons
If the user clicks the form reset button then force the status of the form to be set to clean. The change event is called but we can't use that, or bind reset to checkForm, because the form values are not reset until after the change and reset events have triggered.