Unload event callback
Adding callback for unload event. The callback eanble the developer to
override the default unload behavior. It is useful if you want to
implement other behavior in case of anchor generetaed unload event. for
example - themed modal window - See FaceBook for example. If the user is
Ctrl+F5 the page, the regular unload event is being fired. If the user
is clicking some link - themed modal is being opened. It can be achieved
with AreYouSure using before_unload_callback like the example that I've
attached in the readme.md.
The developer will have to implement the modal and will have to make
sure that the click event will return false.